+++ src/main/java/com/example/green/domain/challenge/controller/dto/admin/AdminTeamChallengeGroupListResponseDto.java
@@ -2,8 +2,8 @@ package com.example.green.domain.challenge.controller.dto.admin;
 
 import java.time.LocalDate;
 
-import com.example.green.domain.challenge.entity.TeamChallengeGroup;
-import com.example.green.domain.challenge.enums.GroupStatus;
+import com.example.green.domain.challenge.entity.group.TeamChallengeGroup;
+import com.example.green.domain.challenge.entity.group.GroupStatus;
 
 import io.swagger.v3.oas.annotations.media.Schema;
