+++ src/main/java/com/example/green/domain/challenge/infra/PersonalChallengeQueryImpl.java
@@ -14,8 +14,8 @@ import com.example.green.domain.challenge.controller.dto.ChallengeDetailDto;
 import com.example.green.domain.challenge.controller.dto.ChallengeListResponseDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminChallengeDetailDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminPersonalChallengesDto;
-import com.example.green.domain.challenge.entity.PersonalChallenge;
-import com.example.green.domain.challenge.enums.ChallengeStatus;
+import com.example.green.domain.challenge.entity.challenge.PersonalChallenge;
+import com.example.green.domain.challenge.entity.challenge.vo.ChallengeStatus;
 import com.example.green.domain.challenge.exception.ChallengeException;
 import com.example.green.domain.challenge.repository.PersonalChallengeRepository;
 import com.example.green.domain.challenge.repository.query.PersonalChallengeQuery;
