+++ src/main/java/com/example/green/domain/challenge/repository/PersonalChallengeParticipationRepository.java
@@ -4,8 +4,8 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.green.domain.challenge.entity.PersonalChallenge;
-import com.example.green.domain.challengecert.entity.PersonalChallengeParticipation;
+import com.example.green.domain.challenge.entity.challenge.PersonalChallenge;
+import com.example.green.domain.challenge.entity.challenge.PersonalChallengeParticipation;
 
 /**
  * 개인 챌린지 참여 정보를 관리하는 레포지토리
