+++ src/main/java/com/example/green/domain/challenge/repository/PersonalChallengeRepository.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import com.example.green.domain.challenge.entity.PersonalChallenge;
+import com.example.green.domain.challenge.entity.challenge.PersonalChallenge;
 
 public interface PersonalChallengeRepository extends JpaRepository<PersonalChallenge, Long> {
