+++ src/main/java/com/example/green/domain/challenge/repository/TeamChallengeGroupRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.example.green.domain.challenge.entity.TeamChallengeGroup;
+import com.example.green.domain.challenge.entity.group.TeamChallengeGroup;
 
 @Repository
 public interface TeamChallengeGroupRepository
