+++ src/main/java/com/example/green/domain/challenge/repository/query/TeamChallengeQuery.java
@@ -7,8 +7,8 @@ import com.example.green.domain.challenge.controller.dto.ChallengeDetailDto;
 import com.example.green.domain.challenge.controller.dto.ChallengeListResponseDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminChallengeDetailDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminTeamChallengesDto;
-import com.example.green.domain.challenge.entity.TeamChallenge;
-import com.example.green.domain.challenge.enums.ChallengeStatus;
+import com.example.green.domain.challenge.entity.challenge.TeamChallenge;
+import com.example.green.domain.challenge.entity.challenge.vo.ChallengeStatus;
 import com.example.green.global.api.page.CursorTemplate;
 import com.example.green.global.api.page.PageTemplate;
