+++ src/main/java/com/example/green/domain/challenge/service/AdminChallengeService.java
@@ -8,13 +8,13 @@ import org.springframework.transaction.annotation.Transactional;
 import com.example.green.domain.challenge.controller.dto.admin.AdminChallengeParticipantListResponseDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminTeamChallengeGroupDetailResponseDto;
 import com.example.green.domain.challenge.controller.dto.admin.AdminTeamChallengeGroupListResponseDto;
-import com.example.green.domain.challenge.entity.TeamChallengeGroup;
+import com.example.green.domain.challenge.entity.group.TeamChallengeGroup;
 import com.example.green.domain.challenge.exception.ChallengeException;
 import com.example.green.domain.challenge.exception.ChallengeExceptionMessage;
 import com.example.green.domain.challenge.repository.PersonalChallengeParticipationRepository;
 import com.example.green.domain.challenge.repository.PersonalChallengeRepository;
 import com.example.green.domain.challenge.repository.TeamChallengeGroupRepository;
-import com.example.green.domain.challengecert.entity.TeamChallengeGroupParticipation;
+import com.example.green.domain.challenge.entity.group.TeamChallengeGroupParticipation;
 import com.example.green.domain.challengecert.repository.TeamChallengeGroupParticipationRepository;
 import com.example.green.domain.challengecert.repository.dao.ChallengeParticipantDao;
 import com.example.green.global.api.page.CursorTemplate;
