+++ src/main/java/com/example/green/domain/challenge/service/TeamChallengeGroupService.java
@@ -9,18 +9,18 @@ import com.example.green.domain.challenge.controller.dto.TeamChallengeGroupCreat
 import com.example.green.domain.challenge.controller.dto.TeamChallengeGroupDetailResponseDto;
 import com.example.green.domain.challenge.controller.dto.TeamChallengeGroupListResponseDto;
 import com.example.green.domain.challenge.controller.dto.TeamChallengeGroupUpdateRequestDto;
-import com.example.green.domain.challenge.entity.TeamChallenge;
-import com.example.green.domain.challenge.entity.TeamChallengeGroup;
-import com.example.green.domain.challenge.entity.vo.GroupAddress;
+import com.example.green.domain.challenge.entity.challenge.TeamChallenge;
+import com.example.green.domain.challenge.entity.group.TeamChallengeGroup;
+import com.example.green.domain.challenge.entity.group.GroupAddress;
 import com.example.green.domain.challenge.exception.ChallengeException;
 import com.example.green.domain.challenge.exception.ChallengeExceptionMessage;
 import com.example.green.domain.challenge.repository.TeamChallengeGroupRepository;
 import com.example.green.domain.challenge.repository.TeamChallengeParticipationRepository;
 import com.example.green.domain.challenge.repository.TeamChallengeRepository;
 import com.example.green.domain.challenge.utils.CodeGenerator;
-import com.example.green.domain.challengecert.entity.TeamChallengeGroupParticipation;
-import com.example.green.domain.challengecert.entity.TeamChallengeParticipation;
-import com.example.green.domain.challengecert.entity.enums.GroupRoleType;
+import com.example.green.domain.challenge.entity.group.TeamChallengeGroupParticipation;
+import com.example.green.domain.challenge.entity.challenge.TeamChallengeParticipation;
+import com.example.green.domain.challenge.entity.group.GroupRoleType;
 import com.example.green.domain.challengecert.repository.TeamChallengeGroupParticipationRepository;
 import com.example.green.domain.member.entity.Member;
 import com.example.green.domain.member.repository.MemberRepository;
