+++ src/main/java/com/example/green/domain/challengecert/repository/TeamChallengeCertificationRepository.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.green.domain.challengecert.entity.TeamChallengeCertification;
-import com.example.green.domain.challengecert.entity.TeamChallengeParticipation;
+import com.example.green.domain.challenge.entity.challenge.TeamChallengeParticipation;
 
 public interface TeamChallengeCertificationRepository
 	extends JpaRepository<TeamChallengeCertification, Long>, TeamChallengeCertificationRepositoryCustom {
